refactor: Moved functions out of `ScatterPlotTab` and into `scatter_plot_data_utils.ts`
(cherry picked from commit 13157e94ab5799d7d3fd45fc8a5a2aa1db5d52b8)
