+++ src/colorizer/viewport/types.ts
@@ -106,6 +106,20 @@ export type RenderCanvasStateParams = {
   interpolate3d: boolean;
 };
 
+/** Subset of RenderCanvasStateParams used for colorization. */
+export const COLORIZE_STATE_KEYS = [
+  "dataset",
+  "featureKey",
+  "colorRamp",
+  "colorRampRange",
+  "categoricalPaletteRamp",
+  "outOfRangeDrawSettings",
+  "outlierDrawSettings",
+  "inRangeLUT",
+] satisfies (keyof RenderCanvasStateParams)[];
+
+export type ColorizeStateParams = Pick<RenderCanvasStateParams, (typeof COLORIZE_STATE_KEYS)[number]>;
+
 export type RenderOptions = {
   /** If true, renders synchronously. */
   synchronous?: boolean;
