+++ src/components/Tabs/ScatterPlot/ScatterPlotTab.tsx
@@ -1,15 +1,16 @@
 import { ExportOutlined } from "@ant-design/icons";
 import { Button, Tooltip } from "antd";
-import Plotly, { type PlotData, type PlotMarker } from "plotly.js-dist-min";
+import Plotly, { type PlotData } from "plotly.js-dist-min";
 import React, { memo, type ReactElement, useCallback, useContext, useEffect, useMemo, useRef, useState } from "react";
 import styled from "styled-components";
-import { Color, type ColorRepresentation } from "three";
+import { Color } from "three";
 
 import { SwitchIconSVG } from "src/assets";
-import { ColorRampType, type Dataset, type Track } from "src/colorizer";
+import { type Dataset } from "src/colorizer";
 import { CENTROID_Y_FEATURE_KEY, TIME_FEATURE_KEY } from "src/colorizer/Dataset";
-import { DrawMode, type HexColorString, PlotRangeType, ViewMode } from "src/colorizer/types";
+import { PlotRangeType, ViewMode } from "src/colorizer/types";
 import { hasAnyValueChanged, isPositiveInteger } from "src/colorizer/utils/data_utils";
+import { ColorizeStateParams } from "src/colorizer/viewport/types";
 import type { ShowAlertBannerCallback } from "src/components/Banner/hooks";
 import IconButton from "src/components/Buttons/IconButton";
 import TextButton from "src/components/Buttons/TextButton";
@@ -23,20 +24,14 @@ import { AppThemeContext } from "src/styles/AppStyle";
 import { FlexRow, FlexRowAlignCenter } from "src/styles/utils";
 import { downloadCsv } from "src/utils/file_io";
 import {
-  type DataArray,
-  getBucketIndex,
+  colorizeScatterplotPoints,
+  filterDataByRange,
   getCrosshairShapes,
   getHistogramBins,
-  getHoverTemplate,
   getScatterplotDataAsCsv,
   isHistogramEvent,
-  makeEmptyTraceData,
   makeLineTrace,
-  scaleColorOpacityByMarkerCount,
   scatterplotTraceToShapes,
-  splitTraceData,
-  subsampleColorRamp,
-  type TraceData,
 } from "src/utils/scatter_plot_data_utils";
 import { areAnyHotkeysPressed } from "src/utils/user_input";
 
@@ -46,12 +41,6 @@ const PLOTLY_CONFIG: Partial<Plotly.Config> = {
   displayModeBar: false,
   responsive: true,
 };
-
-const MAX_POINTS_PER_TRACE = 1024;
-const COLOR_RAMP_SUBSAMPLES = 100;
-const NUM_RESERVED_BUCKETS = 2;
-const BUCKET_INDEX_OUTOFRANGE = 0;
-const BUCKET_INDEX_OUTLIERS = 1;
 
 const DEFAULT_RANGE_TYPE = PlotRangeType.ALL_TIME;
 
@@ -77,7 +66,9 @@ const ScatterPlotContainer = styled.div`
  * A tab that displays an interactive scatter plot between two features in the dataset.
  */
 export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactElement {
-  // ^ Memo prevents re-rendering if the props haven't changed.
+  //////////////////////////////////
+  // State and Refs
+  //////////////////////////////////
   const theme = useContext(AppThemeContext);
 
   const colorRamp = useViewerStateStore((state) => state.colorRamp);
@@ -115,11 +106,36 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
   // Debounce changes to the dataset to prevent noticeably blocking the UI thread with a re-render.
   const datasetKey = useViewerStateStore((state) => state.datasetKey);
   const rawDataset = useViewerStateStore((state) => state.dataset);
-  const rawCategoricalPalette = useViewerStateStore((state) => state.categoricalPalette);
+  const rawCategoricalPaletteRamp = useViewerStateStore((state) => state.categoricalPaletteRamp);
   const rawColorRampRange = useViewerStateStore((state) => state.colorRampRange);
   const dataset = useDebounce(rawDataset, 500);
-  const categoricalPalette = useDebounce(rawCategoricalPalette, 100);
+  const categoricalPaletteRamp = useDebounce(rawCategoricalPaletteRamp, 100);
   const [colorRampMin, colorRampMax] = useDebounce(rawColorRampRange, 100);
+
+  const colorizeConfig = useMemo(
+    () =>
+      ({
+        dataset,
+        featureKey: selectedFeatureKey,
+        colorRamp,
+        colorRampRange: [colorRampMin, colorRampMax],
+        inRangeLUT,
+        categoricalPaletteRamp,
+        outOfRangeDrawSettings,
+        outlierDrawSettings,
+      } satisfies ColorizeStateParams),
+    [
+      dataset,
+      selectedFeatureKey,
+      colorRamp,
+      colorRampMin,
+      colorRampMax,
+      inRangeLUT,
+      categoricalPaletteRamp,
+      outOfRangeDrawSettings,
+      outlierDrawSettings,
+    ]
+  );
 
   const plottedIds = useRef<Set<number>>(new Set());
 
@@ -304,7 +320,7 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
   // Helper Methods
   //////////////////////////////////
 
-  /** Retrieve feature data, if it exists. Accounts for the artificially-added time feature. */
+  /** Retrieve feature data, if it exists. */
   const getData = (featureKey: string | null, dataset: Dataset | null): Uint32Array | Float32Array | undefined => {
     if (featureKey === null || dataset === null) {
       return undefined;
@@ -328,110 +344,6 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
     // TODO: Show histograms on default, cleared layout
     Plotly.react(plotDivRef.current!, [], {}, PLOTLY_CONFIG);
     setIsRendering(false);
-  };
-
-  /**
-   * Removes data from all indices where xData or yData is NaN or Infinity.
-   */
-  const sanitizeNumericDataArrays = (
-    xData: DataArray,
-    yData: DataArray,
-    objectIds: number[],
-    segIds: number[],
-    trackIds: number[]
-  ): { xData: DataArray; yData: DataArray; objectIds: number[]; segIds: number[]; trackIds: number[] } => {
-    // Boolean array, true if both x and y are not NaN/infinity
-    const isFiniteLut = Array.from(Array(xData.length)).map(
-      (_, i) => Number.isFinite(xData[i]) && Number.isFinite(yData[i])
-    );
-
-    return {
-      xData: xData.filter((_, i) => isFiniteLut[i]),
-      yData: yData.filter((_, i) => isFiniteLut[i]),
-      objectIds: objectIds.filter((_, i) => isFiniteLut[i]),
-      segIds: segIds.filter((_, i) => isFiniteLut[i]),
-      trackIds: trackIds.filter((_, i) => isFiniteLut[i]),
-    };
-  };
-
-  /**
-   * Reduces the given data to only show the selected range (frame, track, or
-   * all data points).
-   * @param rawXData raw data for the X-axis feature
-   * @param rawYData raw data for the Y-axis feature.
-   * @param range The range type to filter the data by.
-   * @param track Required if `range` is `PlotRangeType.CURRENT_TRACK`. The
-   * track to filter data by.
-   * @returns One of the following:
-   *   - `undefined` if the data could not be filtered.
-   *   - An object with the following arrays:
-   *     - `xData`: The filtered x data.
-   *     - `yData`: The filtered y data.
-   *     - `objectIds`: The object IDs corresponding to the index of the
-   *       filtered data.
-   */
-  const filterDataByRange = (
-    rawXData: DataArray,
-    rawYData: DataArray,
-    range: PlotRangeType,
-    track?: Track
-  ):
-    | undefined
-    | {
-        xData: DataArray;
-        yData: DataArray;
-        objectIds: number[];
-        segIds: number[];
-        trackIds: number[];
-      } => {
-    if (!dataset || !rawXData || !rawYData) {
-      return undefined;
-    }
-
-    let xData: DataArray = [];
-    let yData: DataArray = [];
-    let objectIds: number[] = [];
-    let segIds: number[] = [];
-    let trackIds: number[] = [];
-
-    if (range === PlotRangeType.CURRENT_FRAME) {
-      // Filter data to only show the current frame.
-      if (!dataset.times) {
-        return undefined;
-      }
-      for (let i = 0; i < dataset.times.length; i++) {
-        if (dataset.times[i] === currentFrame) {
-          objectIds.push(i);
-          segIds.push(dataset.getSegmentationId(i));
-          trackIds.push(dataset.getTrackId(i));
-          xData.push(rawXData[i]);
-          yData.push(rawYData[i]);
-        }
-      }
-    } else if (range === PlotRangeType.CURRENT_TRACK) {
-      // Filter data to only show the current track.
-      if (!track) {
-        return { xData: [], yData: [], objectIds: [], segIds: [], trackIds: [] };
-      }
-      for (let i = 0; i < track.ids.length; i++) {
-        const id = track.ids[i];
-        xData.push(rawXData[id]);
-        yData.push(rawYData[id]);
-      }
-      objectIds = Array.from(track.ids);
-      segIds = objectIds.map(dataset.getSegmentationId);
-      trackIds = Array(track.ids.length).fill(track.trackId);
-    } else {
-      // All time
-      objectIds = [...rawXData.keys()];
-      segIds = objectIds.map(dataset.getSegmentationId);
-      trackIds = Array.from(dataset!.trackIds || []);
-      // Copying the reference is faster than `Array.from()`.
-      xData = rawXData;
-      yData = rawYData;
-    }
-    // TODO: Consider moving this or making it conditional if it causes performance issues.
-    return sanitizeNumericDataArrays(xData, yData, objectIds, segIds, trackIds);
   };
 
   /**
@@ -529,186 +441,6 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
     );
   };
 
-  // TODO: Move to `scatter_plot_data_utils.ts`
-  /**
-   * Applies coloring to point traces in a scatterplot. Does this by splitting
-   * the data into multiple traces each with a solid color, which is much faster
-   * than using Plotly's native color ramping. Also enforces a maximum number of
-   * points per trace, which significantly speeds up Plotly renders.
-   *
-   * @param xData
-   * @param yData
-   * @param objectIds
-   * @param trackIds
-   * @param markerConfig Additional marker configuration to apply to all points.
-   * By default, markers are size 4.
-   * @param {Color | undefined} overrideColor When defined, uses a base color
-   * for all points, instead of calculating based on the color ramp or palette.
-   * @param allowHover Whether to allow hover tooltips on the points, true by
-   * default. When false, hover info is disabled.
-   */
-  const colorizeScatterplotPoints = (
-    xData: DataArray,
-    yData: DataArray,
-    objectIds: number[],
-    segIds: number[],
-    trackIds: number[],
-    markerConfig: Partial<PlotMarker> & { outliers?: Partial<PlotMarker>; outOfRange?: Partial<PlotMarker> } = {},
-    overrideColor?: Color,
-    allowHover = true
-  ): Partial<PlotData>[] => {
-    if (selectedFeatureKey === null || dataset === null || !xAxisFeatureKey || !yAxisFeatureKey) {
-      return [];
-    }
-    const featureData = dataset.getFeatureData(selectedFeatureKey);
-    if (!featureData) {
-      return [];
-    }
-
-    // Generate colors
-    const categories = dataset.getFeatureCategories(selectedFeatureKey);
-    const isCategorical = categories !== null;
-    const isCategoricalRamp = colorRamp.type === ColorRampType.CATEGORICAL;
-    const usingOverrideColor = markerConfig.color || overrideColor;
-    overrideColor = overrideColor || new Color(markerConfig.color as ColorRepresentation);
-
-    let colors: Color[];
-    if (usingOverrideColor) {
-      // Do no coloring! Keep all points in the same bucket, which will still be split up later.
-      colors = [overrideColor];
-    } else if (isCategorical) {
-      colors = categoricalPalette;
-    } else if (isCategoricalRamp) {
-      colors = colorRamp.colorStops;
-    } else {
-      colors = subsampleColorRamp(colorRamp, COLOR_RAMP_SUBSAMPLES);
-    }
-
-    const colorMinValue = isCategorical ? 0 : colorRampMin;
-    const colorMaxValue = isCategorical ? categories.length - 1 : colorRampMax;
-
-    // Make a bucket group for each ramp/palette color and for the out-of-range and outliers.
-    const traceDataBuckets: TraceData[] = [];
-    const overrideColorHex: HexColorString = `#${overrideColor.getHexString()}`;
-
-    let outOfRangeColor: HexColorString = `#${outOfRangeDrawSettings.color.getHexString()}`;
-    let outlierColor: HexColorString = `#${outlierDrawSettings.color.getHexString()}`;
-    const outOfRangeMarker = { ...markerConfig, ...markerConfig.outOfRange };
-    const outlierMarker = { ...markerConfig, ...markerConfig.outliers };
-    if (usingOverrideColor) {
-      outlierColor = overrideColorHex;
-      outOfRangeColor = overrideColorHex;
-    }
-
-    traceDataBuckets.push(makeEmptyTraceData(outOfRangeColor, outOfRangeMarker)); // 0 = out of range
-    traceDataBuckets.push(makeEmptyTraceData(outlierColor, outlierMarker)); // 1 = outliers
-
-    for (let i = NUM_RESERVED_BUCKETS; i < colors.length + NUM_RESERVED_BUCKETS; i++) {
-      let color: HexColorString = `#${colors[i - NUM_RESERVED_BUCKETS].getHexString()}`;
-      const marker = markerConfig;
-      if (usingOverrideColor) {
-        color = overrideColorHex;
-      }
-      traceDataBuckets.push(makeEmptyTraceData(color, marker));
-    }
-
-    // Sort data into buckets
-    for (let i = 0; i < xData.length; i++) {
-      const objectId = objectIds[i];
-      const isMinMaxNaN = Number.isNaN(colorMaxValue) && Number.isNaN(colorMinValue);
-      const isNaN = Number.isNaN(featureData.data[objectId]);
-      const isOutlier = dataset.outliers ? dataset.outliers[objectId] : false;
-      const isOutOfRange = inRangeLUT[objectId] === 0;
-
-      if (Number.isNaN(objectId) || objectId === undefined || objectId <= 0) {
-        continue;
-      }
-
-      let bucketIndex;
-      if (isOutOfRange) {
-        bucketIndex = BUCKET_INDEX_OUTOFRANGE;
-      } else if (isOutlier || isNaN || isMinMaxNaN) {
-        bucketIndex = BUCKET_INDEX_OUTLIERS;
-      } else if (usingOverrideColor) {
-        bucketIndex = NUM_RESERVED_BUCKETS;
-      } else if (isCategorical || isCategoricalRamp) {
-        bucketIndex = (Math.round(featureData.data[objectId]) % colors.length) + NUM_RESERVED_BUCKETS;
-      } else {
-        bucketIndex =
-          getBucketIndex(featureData.data[objectId], colorMinValue, colorMaxValue, colors.length) +
-          NUM_RESERVED_BUCKETS;
-      }
-
-      const bucket = traceDataBuckets[bucketIndex];
-      bucket.x.push(xData[i]);
-      bucket.y.push(yData[i]);
-      bucket.objectIds.push(objectIds[i]);
-      bucket.segIds.push(segIds[i]);
-      bucket.trackIds.push(trackIds[i]);
-    }
-
-    // Apply transparency to the colors
-    const totalPoints = xData.length;
-    const numOutOfRange = traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].x.length;
-    const numOutliers = traceDataBuckets[BUCKET_INDEX_OUTLIERS].x.length;
-    const numInRange = totalPoints - numOutOfRange - numOutliers;
-    // Use total number to calculate transparency for the out of range and outlier buckets, so they do not appear
-    // unusually opaque if there are only a small number of points.
-    traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].color = scaleColorOpacityByMarkerCount(
-      totalPoints,
-      traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].color
-    );
-    traceDataBuckets[BUCKET_INDEX_OUTLIERS].color = scaleColorOpacityByMarkerCount(
-      totalPoints,
-      traceDataBuckets[BUCKET_INDEX_OUTLIERS].color
-    );
-    traceDataBuckets.slice(2).forEach((bucket) => {
-      bucket.color = scaleColorOpacityByMarkerCount(numInRange, bucket.color);
-    });
-
-    // Optionally delete the outlier and out of range buckets to hide the values.
-    if (outlierDrawSettings.mode === DrawMode.HIDE && !markerConfig.outliers) {
-      traceDataBuckets.splice(1, 1);
-    }
-    if (outOfRangeDrawSettings.mode === DrawMode.HIDE && !markerConfig.outOfRange) {
-      traceDataBuckets.splice(0, 1);
-    }
-
-    // Transform buckets into traces
-    const traces: Partial<PlotData>[] = traceDataBuckets
-      .filter((bucket) => bucket.x.length > 0) // Remove empty buckets
-      .reduce((acc: TraceData[], bucket: TraceData) => {
-        // Split the traces into smaller chunks to prevent plotly from freezing.
-        acc.push(...splitTraceData(bucket, MAX_POINTS_PER_TRACE));
-        return acc;
-      }, [])
-      .map((bucket) => {
-        // Custom data is shown in the hover tooltip.
-        // Formatted as [trackId, segId][]
-        const stackedCustomData = bucket.trackIds.map((trackId, index) => {
-          return [trackId.toString(), bucket.segIds[index].toString()];
-        });
-        return {
-          x: bucket.x,
-          y: bucket.y,
-          ids: bucket.objectIds.map((id) => id.toString()),
-          customdata: stackedCustomData,
-          name: "",
-          type: "scattergl",
-          mode: "markers",
-          marker: {
-            color: bucket.color,
-            size: 4,
-            ...bucket.marker,
-          },
-          hoverinfo: allowHover ? "text" : "skip",
-          hovertemplate: allowHover ? getHoverTemplate(dataset, xAxisFeatureKey, yAxisFeatureKey) : undefined,
-        };
-      });
-
-    return traces;
-  };
-
   //////////////////////////////////
   // Plot Rendering
   //////////////////////////////////
@@ -730,7 +462,7 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
     colorRampMax,
     colorRamp,
     inRangeLUT,
-    categoricalPalette,
+    categoricalPaletteRamp,
     resetXAxisPlotRange,
     resetYAxisPlotRange,
   ];
@@ -772,7 +504,7 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
     // Render the dots. See TODO in `scatterplotTraceToShapes` for refactoring
     // `colorizeScatterplotPoints`.
     const pointShapes = scatterplotTraceToShapes(
-      colorizeScatterplotPoints(xData, yData, ids, segIds, trackIds, {
+      colorizeScatterplotPoints(colorizeConfig, xAxisFeatureKey, yAxisFeatureKey, xData, yData, ids, segIds, trackIds, {
         outOfRange: {
           color: theme.color.layout.background,
           line: { width: 2, color: "#" + outOfRangeOutlineColor.getHexString() + "40" },
@@ -795,7 +527,7 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
     }
 
     // Filter data by the range type, if applicable
-    const result = filterDataByRange(rawXData, rawYData, rangeType);
+    const result = filterDataByRange(dataset, currentFrame, rawXData, rawYData, rangeType);
     if (result === undefined) {
       clearPlotAndStopRender();
       return;
@@ -814,6 +546,9 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
 
     // Configure traces
     const traces = colorizeScatterplotPoints(
+      colorizeConfig,
+      xAxisFeatureKey,
+      yAxisFeatureKey,
       xData,
       yData,
       objectIds,
@@ -856,7 +591,14 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
 
     // Render current track as an extra trace.
     for (const track of tracks.values()) {
-      const trackData = filterDataByRange(rawXData, rawYData, PlotRangeType.CURRENT_TRACK, track);
+      const trackData = filterDataByRange(
+        dataset,
+        currentFrame,
+        rawXData,
+        rawYData,
+        PlotRangeType.CURRENT_TRACK,
+        track
+      );
       if (trackData && rangeType !== PlotRangeType.CURRENT_FRAME) {
         // Render an extra trace for lines connecting the points in the current track when time is a feature.
         if (isUsingTime) {
@@ -867,6 +609,9 @@ export default memo(function ScatterPlotTab(props: ScatterPlotTabProps): ReactEl
         // Connect track points as a line trace.
         const outOfRangeOutlineColor = outOfRangeDrawSettings.color.clone().multiplyScalar(0.8);
         const trackTraces = colorizeScatterplotPoints(
+          colorizeConfig,
+          xAxisFeatureKey,
+          yAxisFeatureKey,
           trackData.xData,
           trackData.yData,
           trackData.objectIds,
