+++ src/utils/scatter_plot_data_utils.ts
@@ -1,19 +1,30 @@
 import { unparse } from "papaparse";
 import type { PlotData, PlotMarker, PlotMouseEvent, Shape } from "plotly.js-dist-min";
-import type { Color } from "three";
+import { Color, type ColorRepresentation } from "three";
 
 import {
   type ColorRamp,
+  ColorRampType,
   CSV_COL_FILTERED,
   CSV_COL_OUTLIER,
   CSV_COL_SEG_ID,
   CSV_COL_TIME_WITH_UNITS,
   CSV_COL_TRACK,
   type Dataset,
+  DrawMode,
   type HexColorString,
+  PlotRangeType,
+  Track,
 } from "src/colorizer";
 import { type FeatureData, FeatureType, TIME_FEATURE_KEY, TRACK_FEATURE_KEY } from "src/colorizer/Dataset";
 import { remap } from "src/colorizer/utils/math_utils";
+import { ColorizeStateParams } from "src/colorizer/viewport/types";
+
+const MAX_POINTS_PER_TRACE = 1024;
+const COLOR_RAMP_SUBSAMPLES = 100;
+const NUM_RESERVED_BUCKETS = 2;
+const BUCKET_INDEX_OUTOFRANGE = 0;
+const BUCKET_INDEX_OUTLIERS = 1;
 
 export type DataArray = Uint32Array | Float32Array | number[];
 
@@ -26,6 +37,116 @@ export type TraceData = {
   color: HexColorString;
   marker: Partial<PlotMarker>;
 };
+
+//// Data validation ////
+
+/**
+ * Removes data from all indices where xData or yData is NaN or Infinity.
+ */
+const sanitizeNumericDataArrays = (
+  xData: DataArray,
+  yData: DataArray,
+  objectIds: number[],
+  segIds: number[],
+  trackIds: number[]
+): { xData: DataArray; yData: DataArray; objectIds: number[]; segIds: number[]; trackIds: number[] } => {
+  // Boolean array, true if both x and y are not NaN/infinity
+  const isFiniteLut = Array.from(Array(xData.length)).map(
+    (_, i) => Number.isFinite(xData[i]) && Number.isFinite(yData[i])
+  );
+
+  return {
+    xData: xData.filter((_, i) => isFiniteLut[i]),
+    yData: yData.filter((_, i) => isFiniteLut[i]),
+    objectIds: objectIds.filter((_, i) => isFiniteLut[i]),
+    segIds: segIds.filter((_, i) => isFiniteLut[i]),
+    trackIds: trackIds.filter((_, i) => isFiniteLut[i]),
+  };
+};
+
+/**
+ * Reduces the given data to only show the selected range (frame, track, or
+ * all data points).
+ * @param rawXData raw data for the X-axis feature
+ * @param rawYData raw data for the Y-axis feature.
+ * @param range The range type to filter the data by.
+ * @param track Required if `range` is `PlotRangeType.CURRENT_TRACK`. The
+ * track to filter data by.
+ * @returns One of the following:
+ *   - `undefined` if the data could not be filtered.
+ *   - An object with the following arrays:
+ *     - `xData`: The filtered x data.
+ *     - `yData`: The filtered y data.
+ *     - `objectIds`: The object IDs corresponding to the index of the
+ *       filtered data.
+ */
+export const filterDataByRange = (
+  dataset: Dataset | null,
+  currentFrame: number,
+  rawXData: DataArray,
+  rawYData: DataArray,
+  range: PlotRangeType,
+  track?: Track
+):
+  | undefined
+  | {
+      xData: DataArray;
+      yData: DataArray;
+      objectIds: number[];
+      segIds: number[];
+      trackIds: number[];
+    } => {
+  if (!dataset || !rawXData || !rawYData) {
+    return undefined;
+  }
+
+  let xData: DataArray = [];
+  let yData: DataArray = [];
+  let objectIds: number[] = [];
+  let segIds: number[] = [];
+  let trackIds: number[] = [];
+
+  if (range === PlotRangeType.CURRENT_FRAME) {
+    // Filter data to only show the current frame.
+    if (!dataset.times) {
+      return undefined;
+    }
+    for (let i = 0; i < dataset.times.length; i++) {
+      if (dataset.times[i] === currentFrame) {
+        objectIds.push(i);
+        segIds.push(dataset.getSegmentationId(i));
+        trackIds.push(dataset.getTrackId(i));
+        xData.push(rawXData[i]);
+        yData.push(rawYData[i]);
+      }
+    }
+  } else if (range === PlotRangeType.CURRENT_TRACK) {
+    // Filter data to only show the current track.
+    if (!track) {
+      return { xData: [], yData: [], objectIds: [], segIds: [], trackIds: [] };
+    }
+    for (let i = 0; i < track.ids.length; i++) {
+      const id = track.ids[i];
+      xData.push(rawXData[id]);
+      yData.push(rawYData[id]);
+    }
+    objectIds = Array.from(track.ids);
+    segIds = objectIds.map(dataset.getSegmentationId);
+    trackIds = Array(track.ids.length).fill(track.trackId);
+  } else {
+    // All time
+    objectIds = [...rawXData.keys()];
+    segIds = objectIds.map(dataset.getSegmentationId);
+    trackIds = Array.from(dataset!.trackIds || []);
+    // Copying the reference is faster than `Array.from()`.
+    xData = rawXData;
+    yData = rawYData;
+  }
+  // TODO: Consider moving this or making it conditional if it causes performance issues.
+  return sanitizeNumericDataArrays(xData, yData, objectIds, segIds, trackIds);
+};
+
+//// Plotly utilities ////
 
 /**
  * Sample a color ramp at evenly-spaced points, returning the resulting array of colors.
@@ -40,6 +161,214 @@ export function subsampleColorRamp(colorRamp: ColorRamp, numColors: number): Col
   }
   return colors;
 }
+
+/**
+ * Appends alpha opacity information to a hex color string, making it less opaque as the number of markers increases.
+ */
+export function scaleColorOpacityByMarkerCount(numMarkers: number, baseColor: HexColorString): HexColorString {
+  if (baseColor.length !== 7) {
+    throw new Error("ScatterPlotTab.getMarkerColor: Base color '" + baseColor + "' must be 7-character hex string.");
+  }
+  // Interpolate linearly between 80% and 25% transparency from 0 up to a max of 1000 markers.
+  const opacity = remap(numMarkers, 0, 1000, 0.8, 0.25);
+  const opacityString = Math.floor(opacity * 255)
+    .toString(16)
+    .padStart(2, "0");
+  return (baseColor + opacityString) as HexColorString;
+}
+
+// TODO: Move to `scatter_plot_data_utils.ts`
+/**
+ * Applies coloring to point traces in a scatterplot. Does this by splitting
+ * the data into multiple traces each with a solid color, which is much faster
+ * than using Plotly's native color ramping. Also enforces a maximum number of
+ * points per trace, which significantly speeds up Plotly renders.
+ *
+ * @param xData
+ * @param yData
+ * @param objectIds
+ * @param trackIds
+ * @param markerConfig Additional marker configuration to apply to all points.
+ * By default, markers are size 4.
+ * @param {Color | undefined} overrideColor When defined, uses a base color
+ * for all points, instead of calculating based on the color ramp or palette.
+ * @param allowHover Whether to allow hover tooltips on the points, true by
+ * default. When false, hover info is disabled.
+ */
+export const colorizeScatterplotPoints = (
+  colorizeState: ColorizeStateParams,
+  xAxisFeatureKey: string | null,
+  yAxisFeatureKey: string | null,
+  xData: DataArray,
+  yData: DataArray,
+  objectIds: number[],
+  segIds: number[],
+  trackIds: number[],
+  markerConfig: Partial<PlotMarker> & { outliers?: Partial<PlotMarker>; outOfRange?: Partial<PlotMarker> } = {},
+  overrideColor?: Color,
+  allowHover = true
+): Partial<PlotData>[] => {
+  const {
+    dataset,
+    featureKey,
+    colorRamp,
+    colorRampRange,
+    categoricalPaletteRamp,
+    outOfRangeDrawSettings,
+    outlierDrawSettings,
+    inRangeLUT,
+  } = colorizeState;
+  console.log(colorizeState);
+  if (featureKey === null || dataset === null || !xAxisFeatureKey || !yAxisFeatureKey) {
+    return [];
+  }
+  const featureData = dataset.getFeatureData(featureKey);
+  if (!featureData) {
+    return [];
+  }
+
+  // Generate colors
+  const categories = dataset.getFeatureCategories(featureKey);
+  const isCategorical = categories !== null;
+  const isCategoricalRamp = colorRamp.type === ColorRampType.CATEGORICAL;
+  const usingOverrideColor = markerConfig.color || overrideColor;
+  overrideColor = overrideColor || new Color(markerConfig.color as ColorRepresentation);
+
+  let colors: Color[];
+  if (usingOverrideColor) {
+    // Do no coloring! Keep all points in the same bucket, which will still be split up later.
+    colors = [overrideColor];
+  } else if (isCategorical) {
+    colors = categoricalPaletteRamp.colorStops;
+  } else if (isCategoricalRamp) {
+    colors = colorRamp.colorStops;
+  } else {
+    colors = subsampleColorRamp(colorRamp, COLOR_RAMP_SUBSAMPLES);
+  }
+
+  const colorMinValue = isCategorical ? 0 : colorRampRange[0];
+  const colorMaxValue = isCategorical ? categories.length - 1 : colorRampRange[1];
+
+  // Make a bucket group for each ramp/palette color and for the out-of-range and outliers.
+  const traceDataBuckets: TraceData[] = [];
+  const overrideColorHex: HexColorString = `#${overrideColor.getHexString()}`;
+
+  let outOfRangeColor: HexColorString = `#${outOfRangeDrawSettings.color.getHexString()}`;
+  let outlierColor: HexColorString = `#${outlierDrawSettings.color.getHexString()}`;
+  const outOfRangeMarker = { ...markerConfig, ...markerConfig.outOfRange };
+  const outlierMarker = { ...markerConfig, ...markerConfig.outliers };
+  if (usingOverrideColor) {
+    outlierColor = overrideColorHex;
+    outOfRangeColor = overrideColorHex;
+  }
+
+  traceDataBuckets.push(makeEmptyTraceData(outOfRangeColor, outOfRangeMarker)); // 0 = out of range
+  traceDataBuckets.push(makeEmptyTraceData(outlierColor, outlierMarker)); // 1 = outliers
+
+  for (let i = NUM_RESERVED_BUCKETS; i < colors.length + NUM_RESERVED_BUCKETS; i++) {
+    let color: HexColorString = `#${colors[i - NUM_RESERVED_BUCKETS].getHexString()}`;
+    const marker = markerConfig;
+    if (usingOverrideColor) {
+      color = overrideColorHex;
+    }
+    traceDataBuckets.push(makeEmptyTraceData(color, marker));
+  }
+
+  // Sort data into buckets
+  for (let i = 0; i < xData.length; i++) {
+    const objectId = objectIds[i];
+    const isMinMaxNaN = Number.isNaN(colorMaxValue) && Number.isNaN(colorMinValue);
+    const isNaN = Number.isNaN(featureData.data[objectId]);
+    const isOutlier = dataset.outliers ? dataset.outliers[objectId] : false;
+    const isOutOfRange = inRangeLUT[objectId] === 0;
+
+    if (Number.isNaN(objectId) || objectId === undefined || objectId <= 0) {
+      continue;
+    }
+
+    let bucketIndex;
+    if (isOutOfRange) {
+      bucketIndex = BUCKET_INDEX_OUTOFRANGE;
+    } else if (isOutlier || isNaN || isMinMaxNaN) {
+      bucketIndex = BUCKET_INDEX_OUTLIERS;
+    } else if (usingOverrideColor) {
+      bucketIndex = NUM_RESERVED_BUCKETS;
+    } else if (isCategorical || isCategoricalRamp) {
+      bucketIndex = (Math.round(featureData.data[objectId]) % colors.length) + NUM_RESERVED_BUCKETS;
+    } else {
+      bucketIndex =
+        getBucketIndex(featureData.data[objectId], colorMinValue, colorMaxValue, colors.length) + NUM_RESERVED_BUCKETS;
+    }
+
+    const bucket = traceDataBuckets[bucketIndex];
+    bucket.x.push(xData[i]);
+    bucket.y.push(yData[i]);
+    bucket.objectIds.push(objectIds[i]);
+    bucket.segIds.push(segIds[i]);
+    bucket.trackIds.push(trackIds[i]);
+  }
+
+  // Apply transparency to the colors
+  const totalPoints = xData.length;
+  const numOutOfRange = traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].x.length;
+  const numOutliers = traceDataBuckets[BUCKET_INDEX_OUTLIERS].x.length;
+  const numInRange = totalPoints - numOutOfRange - numOutliers;
+  // Use total number to calculate transparency for the out of range and outlier buckets, so they do not appear
+  // unusually opaque if there are only a small number of points.
+  traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].color = scaleColorOpacityByMarkerCount(
+    totalPoints,
+    traceDataBuckets[BUCKET_INDEX_OUTOFRANGE].color
+  );
+  traceDataBuckets[BUCKET_INDEX_OUTLIERS].color = scaleColorOpacityByMarkerCount(
+    totalPoints,
+    traceDataBuckets[BUCKET_INDEX_OUTLIERS].color
+  );
+  traceDataBuckets.slice(2).forEach((bucket) => {
+    bucket.color = scaleColorOpacityByMarkerCount(numInRange, bucket.color);
+  });
+
+  // Optionally delete the outlier and out of range buckets to hide the values.
+  if (outlierDrawSettings.mode === DrawMode.HIDE && !markerConfig.outliers) {
+    traceDataBuckets.splice(1, 1);
+  }
+  if (outOfRangeDrawSettings.mode === DrawMode.HIDE && !markerConfig.outOfRange) {
+    traceDataBuckets.splice(0, 1);
+  }
+
+  // Transform buckets into traces
+  const traces: Partial<PlotData>[] = traceDataBuckets
+    .filter((bucket) => bucket.x.length > 0) // Remove empty buckets
+    .reduce((acc: TraceData[], bucket: TraceData) => {
+      // Split the traces into smaller chunks to prevent plotly from freezing.
+      acc.push(...splitTraceData(bucket, MAX_POINTS_PER_TRACE));
+      return acc;
+    }, [])
+    .map((bucket) => {
+      // Custom data is shown in the hover tooltip.
+      // Formatted as [trackId, segId][]
+      const stackedCustomData = bucket.trackIds.map((trackId, index) => {
+        return [trackId.toString(), bucket.segIds[index].toString()];
+      });
+      return {
+        x: bucket.x,
+        y: bucket.y,
+        ids: bucket.objectIds.map((id) => id.toString()),
+        customdata: stackedCustomData,
+        name: "",
+        type: "scattergl",
+        mode: "markers",
+        marker: {
+          color: bucket.color,
+          size: 4,
+          ...bucket.marker,
+        },
+        hoverinfo: allowHover ? "text" : "skip",
+        hovertemplate: allowHover ? getHoverTemplate(dataset, xAxisFeatureKey, yAxisFeatureKey) : undefined,
+      };
+    });
+
+  return traces;
+};
 
 /**
  * Returns the index of a bucket that a value should be sorted into, based on a provided range and number of buckets.
@@ -115,21 +444,6 @@ export function splitTraceData(traceData: TraceData, maxPoints: number): TraceDa
  */
 export function isHistogramEvent(eventData: PlotMouseEvent): boolean {
   return eventData.points.length > 0 && eventData.points[0].data.type === "histogram";
-}
-
-/**
- * Appends alpha opacity information to a hex color string, making it less opaque as the number of markers increases.
- */
-export function scaleColorOpacityByMarkerCount(numMarkers: number, baseColor: HexColorString): HexColorString {
-  if (baseColor.length !== 7) {
-    throw new Error("ScatterPlotTab.getMarkerColor: Base color '" + baseColor + "' must be 7-character hex string.");
-  }
-  // Interpolate linearly between 80% and 25% transparency from 0 up to a max of 1000 markers.
-  const opacity = remap(numMarkers, 0, 1000, 0.8, 0.25);
-  const opacityString = Math.floor(opacity * 255)
-    .toString(16)
-    .padStart(2, "0");
-  return (baseColor + opacityString) as HexColorString;
 }
 
 /**
@@ -253,6 +567,8 @@ function isValueOutOfRange(value: number, range?: [number, number]): boolean {
   }
   return false;
 }
+
+//// CSV export utilities ////
 
 export function getScatterplotDataAsCsv(
   dataset: Dataset,
